Pull alias_map code out of internal_rep.rs
